+++ DictinaryTrainer.Infrastructure/Services/Abstract/IWordTrainerService.cs
@@ -5,11 +5,17 @@ namespace DictinaryTrainer.BusinessLogic.Services.Abstract
     public interface IWordTrainerService
     {
         void AddWord(WordDto word);
+
         void DeleteCurrentWord();
+
         WordDto? GetCurrentWord();
+
         WordDto? GetNewWord();
+
         int LoadAllWords();
-        void SetUser(short userId);
+
+        void SetUser(int userId);
+
         void UpdateCurrentWord();
     }
 }
